refactor(module): FileHandler를 infrastructure 모듈로 이동
