+++ stempo-application/src/main/java/com/stempo/service/FileService.java
@@ -5,10 +5,10 @@ import com.stempo.dto.PagedResponseDto;
 import com.stempo.dto.request.DeleteFileRequestDto;
 import com.stempo.dto.response.UploadedFileResponseDto;
 import com.stempo.exception.NotFoundException;
-import com.stempo.handler.FileHandler;
 import com.stempo.mapper.UploadedFileDtoMapper;
 import com.stempo.model.UploadedFile;
 import com.stempo.util.EncryptionUtils;
+import com.stempo.util.FileHandler;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
