+++ stempo-infrastructure/src/main/java/com/stempo/util/FileHandler.java
@@ -1,7 +1,5 @@
-package com.stempo.handler;
+package com.stempo.util;
 
-import com.stempo.util.FileUtils;
-import com.stempo.util.LogSanitizerUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FilenameUtils;
 import org.springframework.beans.factory.annotation.Value;
